+++ client/src/main/java/Main.java
@@ -1,4 +1,7 @@
 import ui.*;
+import ui.menu.InvalidMenuCommandHandler;
+import ui.menu.MenuCommandHandler;
+import ui.menu.MenuCommandHandlerFactory;
 import util.SerializerFactory;
 import java.util.logging.*;
