+++ client/src/main/java/ui/menu/LogoutUserCommandHandler.java
@@ -1,4 +1,7 @@
-package ui;
+package ui.menu;
+
+import ui.AppState;
+import ui.ServerFacade;
 
 import java.io.IOException;
 import java.net.ConnectException;
