+++ client/src/main/java/ui/menu/MenuCommandHandler.java
@@ -1,4 +1,4 @@
-package ui;
+package ui.menu;
 
 public interface MenuCommandHandler {
     String execute(String... arguments);
