+++ client/src/main/java/ui/menu/RegisterUserCommandHandler.java
@@ -1,6 +1,9 @@
-package ui;
+package ui.menu;
 
 import model.SessionData;
+import ui.AppState;
+import ui.ServerFacade;
+
 import java.io.IOException;
 import java.net.ConnectException;
 import java.util.logging.Level;
